+++ frontend/src/shared/components/Toast/hooks/useToast.ts
@@ -1,5 +1,4 @@
 import { useSyncExternalStore } from "react";
-import { getDistributedToasts } from "../store/toastActions";
 import { toastStore } from "../store/toastStore";
 
 export function useToast() {
@@ -8,17 +7,4 @@ export function useToast() {
     () => toastStore.getState(),
     () => toastStore.getState()
   );
-}
-
-export function useDistributedToasts() {
-  const { toasts, defaultPosition, limit } = useToast();
-  const { filteredToasts, queue } = getDistributedToasts({ toasts, defaultPosition, limit });
-
-  return {
-    allToasts: toasts,
-    filteredToasts,
-    queue,
-    defaultPosition: defaultPosition,
-    limit: limit,
-  };
 }
