+++ frontend/src/shared/components/Toast/toast.ts
@@ -1,10 +1,6 @@
 import { showToast } from "./store/toastActions";
 import type { ToastCustomOptions } from "./types/toast.type";
 
-export { useDistributedToasts, useToast } from "./hooks/useToast";
-
-export { toastStore } from "./store/toastStore";
-
 export const toast = {
   success: (message: string, options?: ToastCustomOptions) => showToast({ type: "success", message, ...options }),
